CLOTA: Fix Javacrash on PauseActivity
BZ: 134526

Javacrash occurred on PhoneDoctor due to a unregistration of a broadcast
receiver not registered. Last fix done on this issue added a check to test if
the receiver is connected or not before to unregistrer/register it. The problem
is that the unregister/register process is not synchronized and it's possible
that Service and Activity both finish and call in same time the unregister
function. In this condition, the receiver will be unregistered in a first time and
the next call to unregister will throw an IllegalArgumentException.

To solve the problem unregisterReceiver and registerReceiver were become
synchronized.

Change-Id: I3cd4bdd3f891c237bb0253e2b81767e326113614
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/128590
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>