CLOTA: Consolidate IFW version
BZ: 121954

ifwi version, iafw version, scufw version, punit version or valhooks version
could have bad value and in this case the device build signature could be
wrong. When a signature is corrupted, CrashTool can't link an event to its
weekly. Some crashes could be missed in CrashTool metrics.

To solve this issue, a check was added on the Build creation to test if
ifwi, iafw, scufw, punit and valhooks version are good. If not, the build
signature is replaced by the build signature of the last event inserted in
the database.

Change-Id: I3278ca825481a51066db0125ccfa8883343eb420
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/128842
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Guerard, Christophe <christophe.guerard@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>