CLOTA: Phone Doctor doesn't automatically report events
BZ: 135278

In 'always report' upload mode, PhoneDoctor reports events without notifying
the user. Each time an event occurs it is uploaded immediatly.
When the user launches StartServiceActivity in the same time that an event is
beeing uploaded the askForUpload dialog is displayed. Then the choice selected
by the user is saved and replace the old one(always report). In this way,
in the futur, when an event will occur the user will have to upload it manually
or change again the upload mode.

This is due to a malfunction with the InternalWifi mode. Indeed, if wifiOnly
is set to false (always) the askForUpload message is sent to
StartServiceActivity and wait for an user response.
This behavior was changed and now depends on the upload report management.

Change-Id: I2fd1b7ea9131991bce7d61027dca90c7e2ba7286
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/129090
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>