CLOTA: Add an event filter in PhoneDoctor
BZ: 139523

An filter was added in StartServiceActivity.
This allows the user to select which kind of events he wants to display in
the events list. (All events, all events without STATS events, CRASH events
only).

The Critical crashes notification and crashes notification were modified.
Indeed, when the user clicks on the notification, the StartSerciceActivity
is now displayed with the events filter set to 'CRASH only'.

Change-Id: I83e19e0c9893b8e7f7a38bf77f798592eec48800
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/130579
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>