CLOTA: ADD RESETSRC0 and RESETSRC1 values to DATA0 and DATA1
BZ: 135243

RESETSRC0 and RESETSRC1 (or RESETIRQ1 and RESETIRQ2, depending on the CPU
board) are now copied to /sys/kernel/debug/intel_scu_osnib. The content of
each files is copied in the eventfile of the REBOOT event under DATA3 and
DATA4.

When a REBOOT event occurs, PhoneDoctor reads the eventfile's content and get
the DATA3 and DATA4 set in it. The DATA3 and DATA4 of the REBOOT event are
filled with the read value.
In case of not ready value (when the RESET files are not present), a timer is
launched. When the timer over or if the crashlogger notifies PhoneDoctor,
DATA3 and DATA4 are read again.

Change-Id: I6d2cafb1139e096263f6cf3044a997ec4f74ee6d
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/132618
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>