LIBPARSING: Add Data4 and Data5 in crashfile generation
BZ: 138438

MMGR sends a message to crashlogger when a AP-Reset is done. A new API has been
implemented in which the reset causes are given in the MMGR message.
If reset causes exist an APIMR event is created with data0 set to AP-Reset
origin and data1 to data5 set to cause number 1 to 5 if they exist.

Crashparsing was modified to add data4 and data5 in the crashfile.

Change-Id: I9a457f49c87c85039342ce97095c5ea0081adddd
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/133310
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>