CRASHPARSING: Manage JAVACRASH linked to a TOMBSTONE
BZ: 140415

Since creation of MR2, TOMBSTONE due to a Java world failure generate
a JAVACRASH. This JAVACRASH is actually not parsed by PhoneDoctor and Data0,1,2
are empty.

This is due to the fact that this new kind of JAVACRASH are computed by an
eventfile that is not caught by crashparsing.
Crashparsing was modified to parse the eventfile associated to this JAVACRASH.
In case of native JAVACRASH data0 is filled with the process which causes
the TOMBSTONE and data1 is filled with TOMBSTONE.

Change-Id: I62728c2f83e559cab6b3503f06f4e0e4977a5592
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/133766
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>