CLOTA: Psi phone doctor keeps running
BZ: 141117

When data logs are not available (wrong right access, wrong name, not present)
PhoneDoctor keeps trying to upload them and never stops.
This is due to the fact that if data are not available the upload is skipping
and go to the next event to upload. If there is no other event PhoneDoctor will
try to upload again the current event.

To solve the problem, a new file unvailableFiles was created under each event
crashdir. Each log/data file name of data not available is stored in this
dedicated file. Data not available are not sent and unvailableFiles is sent
instead.

Change-Id: I7fb93042c4d1569a06055d3b1e7fe9f8ff3b8c08
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/135874
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>