CLOTA: Manage fields with = in crashparsing
BZ: 141288

Some crashfile could contain fields with '=' inside. For example an apimr could
contain '=' in one of this reset cause. PhoneDoctor missparses this kind
of value and the '=' and text behind are not stored in the parsed data.

This is due to a bad use of the Java String Splitter.
Indeed, in MainParser.java all crashfile lines are splitted in using
'=' separator but only the first and second occurences are used to get
the parameter name and its value.
simpleAwk function was modified to split crashfile lines in n parts.
The first part is used to get the parameter name and all parts left are used
to get the value.

Change-Id: I909ac6bf039de56c32f21a8d9100595f034228aa
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/135906
Reviewed-by: Rocher, Jeremy <jeremy.rocher@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>