CLOTA: Fix NullPointerException on PhoneDoctor
BZ: 143029

There are several NullPointerException which occur on PhoneDoctor.
-Some of them are due to the fact that getters are used to retrieve objects.
In If conditions the null value is test first and then functions are called
on the object retrieved in still using the getter function. The getter could
fail the second time and that generates a NullPointerException.
-Sometimes database access fail. This is due to the fact that null value on
database object is not tested. Moreover, in close function of GeneralEventDB,
the close method of SQLOpenHelper is called which closes all database access
opened. In case where several threads have an access to the database, if one
of them close its access all other access will be closed too.

In GeneralEventDB the close method was modified.
Temporary object was added to avoid to use getters too many times.

Change-Id: Ib73e4d877c153f0364318be3509944601c34adb7
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>