CLOTA: Fix issue on CrashReportService
BZ: 139086

When CrashReportService is stopped, onKillService() is called. This is a method
of StartServiceActivity which hides all dialogs displayed on the screen and
disable all elements of StartServiceActivity. A Javacrash could occur during
this process because onKill() is called outside of StartServiceActivity thread
and manipulates views outside of their context.

To solve the problem a new message was added in StartServiceActivity. Each time
CrashReportService is killed, this message is sent to StartServiceActivity
in order to call onKillService.

A modification in GeneralEventDB was done to change an SQL request.
In fetchEventsLogsNotUploaded, the logs directory wasn't checked and all events
which have uploaded = 1 and logsuploaded = 0 were taken in account in
the request result.

Change-Id: Ief06e7a9686c30c68d7ae61ac432a82af07d199a
Signed-off-by: Vidoine, Charles-EdouardX <charles.edouardx.vidoine@intel.com>
Reviewed-on: http://android.intel.com:8080/136641
Reviewed-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Reviewed-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Tested-by: Imougar, JacquesX <jacquesx.imougar@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>