PhoneDoctor: remove error message on build_properties.json
BZ: 157525

Currently when the file build_properties.json is not found on device, an error
message is written to logcat.
As this file is considered optional, we want to remove the error message when
it is not found.

When PhoneDoctor is run on a device where no build_properties.json file is
present, no error message (Java exception) shall be displayed.

Change-Id: I86280362c22eabbc785246c99bb79215e600c705
Signed-off-by: Sebbane, AdrienX <adrienx.sebbane@intel.com>