crashreport : Update full DropBox detection on phone doctor
DropBoxManager.isFull was introduced by Intel with BZ 58848. If the
patch is not present then there is no way of checking Full Dropbox
from PhoneDoctor in the previous implementation.
Present implementation wants to use the standart api to determine if
a file was added or dropped from the DropBox, by checking the flags
of the last added item.

Change-Id: I19bd29ef9888cad5bbbd8699714216942f3e8d69
Tracked-On: IMINAN-3286
Signed-off-by: Nicolae Natea <nicolaex.natea@intel.com>
Reviewed-on: 279624