crashreport : [Battery Drain] Phone gets warm with little usage/idle
BZ: 207296

aplog analysis showed that more upload threads were ran
simultaneously. code analysis revealed that if two intents to
crashreportservice were received, close to each other, the
second one could stop the service which could in turn lead to
leaving an upload thread hanging, since the upload thread was
stoped only in 3 cases: protocol error, user request, no internet
connection. after this a new upload thread could get started,
which will first fetch all the entries to be uploaded. So the
threads would work on the same data.
Additionally, a fix was made for RainEvents because while
creating the where clause clause for an sqlite querry, the escape
characters are not taken into account and this could lead to a
sqlite syntax error.

Change-Id: I4bc9d3d061b2e925bb1abed7d195447aa5b44aa3