crashreport: CheckEventsService's could 'exit' before processing
CheckEventsService's could 'exit' before processing of a prior
intent was finished, leading to a thread running in the background
while another one that would perform the same task to stop it's
activity leaving room for another intent to restart the cycle.
Issue should be mainly extra processing time. Similar to BZ:207296

Change-Id: I76eed95fb09afb149581915414b54f680ae2a88e
Tracked-On: IMINAN-3303
Signed-off-by: Nicolae Natea <nicolaex.natea@intel.com>
Reviewed-on: 279661