phonedoctor : add dataready to MPANIC
PD waits for sending files (aplog, core dump, ...) related to MPANIC
but sends the header of the event.

The files will remain on the board, until modemversion is available.
In this case, a file is added with the appropriate modem version.
This modem version will be used by the parser.

A reboot always cleans dataready states.
Fix kw and patch is disabled if modem extension is detected.

Change-Id: I4e94ec5f5f23c0100686e4af2e0a4481696ac9df
Tracked-On: IMINAN-8991
Signed-off-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Signed-off-by: Nicolae Natea <nicolaex.natea@intel.com>
Reviewed-on: 300297