crashreport: update PDStatus for non-saltbay variant
When Phone Doctor status is computed, we sometimes see "Z" value
for variant checks on non-Saltbay devices.
This case occurs when modem version check is not activated
(or if an error occurs when modem_version.txt is generated).
We want to allow the modem version to be empty on some devices
and have a "0" value in that case.

Change-Id: I63eaa1ccae50313b4a205610aed3d9d016343bb1
Tracked-On: IMINAN-6347
Signed-off-by: Nicolae Natea <nicolaex.natea@intel.com>
Reviewed-on: 291423