crashreport: clear user fields  after reporting a bug
Phone doctor bug report page needs fields reset after reporting.
The bug appears when the user tries to re-open an activity before
the async task launched from the activity finishes.

Proposed solution:
If the user tries to open the "report a bug" activity before values
get commited to a file(which is done through an async task) then
he will be prompted with a pop-up asking him to wait a bit longer
otherwise, the activity will be opened with the previous values
still filled. As a fallback mechanism, this prompt is showed only
a defined number of times. If the user does not want to wait, the
activity will be loaded with the fields likely still filled.

Change-Id: If10e2c5cc9cafca96eb6aafcab53f5d2b489a560
Tracked-On: IMINAN-6346
Signed-off-by: Nicolae Natea <nicolaex.natea@intel.com>
Reviewed-on: 291671