crashreport: update crashparsing and rain of crashes mechanism
In the case of tombstones, crashparsing was not picking the signal
that was generated and data1 was not filled in the crashfile.
In the case of rain of crashes, if the fault address cannot be picked
(Data3) use also the stack frame/libs loaded(Data2).
Remove CrashSignature class as it seems redundant with RainSignature.

Change-Id: Ic24ab80dcc1136e17fe6f9774de67dd2611589db
Tracked-On: IMINAN-22677
Signed-off-by: Nicolae Natea <nicolaex.natea@intel.com>
Reviewed-on: 334080