crashreport: monitor new ingredients section
A new ingredients section, GETBULKPROP, was added to ingredients.conf
that should be checked and sent to crashtool.
Additional processing is done on the new entries to format the
output in a more friendly manner for crashtool.
JsonObject is now used for maintaining the list of ingredients
instead of a HashMap.

Change-Id: Ied633c6f0239dfa4407d4bf8b29203601ca55f83
Tracked-On: GMINL-9966
Signed-off-by: Nicolae Natea <nicolaex.natea@intel.com>
Reviewed-on: 392523