crashreport: add service to manage free space on the logs partition
Phone Doctor crash logs are often large, and especially on devices
with a low amount of available storage this can become an issue very
quickly as the disk gets full.

Proposed solution is to have a more intrusive space management service that
will check available space on the logs partition. In case the disk is full
over a certain threshold, the uploaded logs are removed. As a fallback,
the oldest logs will be removed until a certain threshold is reached, if
removing the uploaded logs will not suffice.

Change-Id: Iaef45ff623a9d5c28f7432bf8f10e6db03f12110
Tracked-On: GMINL-9546
Signed-off-by: Nicolae Natea <nicolaex.natea@intel.com>
Reviewed-on: 372521
Reviewed-by: Kumar-mayernik, Nisha <nisha.kumar-mayernik@intel.com>
Tested-by: Kumar-mayernik, Nisha <nisha.kumar-mayernik@intel.com>