log_infra: Update all modules output path
Modules output has been changed from system/... to system/vendor/...
As the modules are part of a project located in vendor/intel/ in the
 source tree. The path for the file generated by such module shall
 be in system/vendor/.
This is for better isolation of Intel binaries and modifications.
Crashparsing and com.google.gson are built as static libraries,
linked to crashreport to avoid adding extra permissions.

With this filesystem organization it is possible to enable the
 generation of a vendor partition.

Change-Id: I8fc70c7af7dcd40d253beab0b74fbcfc9dcf250a
Tracked-On: OAM-1147
Signed-off-by: Benjamin Pujol <benjamin.pujol@intel.com>
Signed-off-by: Nicolae Natea <nicolaex.natea@intel.com>
Reviewed-on: 401283