crashreport: replace logs folder with a txt file, on cleanup
when low on space, and removing old paths, remove crashfolder.
In order to have the reason for removal, add in its place a
file with the pattern: Cleaned_By_<issuer>_<dateOfCleaning>.txt

Mechanism to clean unreferenced paths has been updated to
have a delay between the time that a path was modified and
the actual cleanup, so that, if PhoneInspector runs before
crashreport acknowledged a new event, the unreferenced path
will not be cleaned, unless a defined period of time has passed.

Change-Id: Idabdd81f27cd3b7f629e589f2e4b9dd445fff7c3
Tracked-On: OAM-5290
Signed-off-by: Nicolae Natea <nicolaex.natea@intel.com>
Reviewed-on: 427020