crashreport: start cleanup mechanism according to user defined quota
A user defined quota will be taken into account when managing
free space on the target partition. The user will be able to
change the quota by setting a property:
persist.crashlogd.data_quota.
This, in turn, changing the time in which the crash cleanup
management mechanism gets triggered, and the time in which
crashlog stop generating crashes.

Change-Id: If19bf232ab213fc6d86433a741397c8399b20ee8
Tracked-On: OAM-6514
Signed-off-by: Nicolae Natea <nicolaex.natea@intel.com>
Reviewed-on: 433900