crashparsing: adapt MPANIC parsing
For modem silent reset case, MPANIC parsing is
updated to match zipped data file.

Also, event crashdata management should be updated to
handle both data from parsing & crashlog.

It means that if data is already present in crashfile (crashlog source)
and not computed by parsing (phone doctor source), it will be preserved
in the final result.

This was not the case before, only phone doctor source was preserved.

Change-Id: Idd0010b8fdf849efeca691188176efc22a1f1bc8
Signed-off-by: Benoit, NicolasX <nicolasx.benoit@intel.com>
Tracked-On: OAM-20180
Reviewed-on: 466672