crashreport: add disclaimer before android setup wizard
A disclaimer will be displayed on phone start. This disclaimer
will appear after the first boot and after a factory reset. The
puropse of the disclaimer is to only inform the user that data
collection is taking place. It will be displayed after the
google setup wizard.
It also implements a similar bypass mechanism to google setup
wizard. However the mechanism monitors only for 4 consecutive
taps in the corner of the device screen, regardless of order.
Other taps would cancel the skip mechanism. If this mechanism
is activated, google setup wizard is also bypassed. The back
action closes the activity.

Change-Id: Iee8072a66079989d6b7f39474c6a9c6aa52faed4
Tracked-On: OAM-19135
Signed-off-by: Nicolae Natea <nicolaex.natea@intel.com>
Reviewed-on: 470464