crashreport: do not remove original file if compression is deemed failed
Upon successful compression of large logfiles, the original
files are removed.
However, we've seen events that have empty archives, which
would suggest that the compression appeared as successful,
and therefore permitted the removal of the original file.
According to the documentation the FileInputStream implementation
should block if the passed buffer is valid, until some data could
be read or the end of file is reached. There is a scenario however
in which the native read function could return EAGAIN which would
cause the FileInputStream implementation to report the read size
as 0. This scenario does not seem plausible, but the code was
updated with this in mind.
An extra check is added to avoid the removal of the original file
if the size of the compressed file is 0.

Change-Id: I3d4b10159011c4a84a07525318a73be091f33acb
Tracked-On: OAM-14215
Signed-off-by: Nicolae Natea <nicolaex.natea@intel.com>
Reviewed-on: 471046