crashreport: add an option to disable UI notification of crashes
Crash notification should be gated by a config parameter:
"enable_crash_notification". If the config is set to true, there will
be no change to the current behaviour, otherwise, no crash notification
should appear and the reference to this feature from the Settings menu
shall be hidden. By default the config is set to true.

Change-Id: I1260f33b059270fbba80e5cc3aeffd0068ce66e0
Tracked-On: OAM-22201
Signed-off-by: Nicolae Natea <nicolaex.natea@intel.com>
Reviewed-on: 471560