[REVERT ME] priv-app in vendor depends on AOSP patches
Privileged applications put in vendor will not be found.

The two following patches, already merged on Google master branch
are required:
https://android-review.googlesource.com/#/c/177621/
https://android-review.googlesource.com/#/c/182240/

But they are not present yet in the current AOSP release.

Goal is to keep following Google's direction to move vendor specific
binaries to vendor.
Hence check if the priv-app in vendor are supported.

This patch will be reverted later when two patches below are merged:
https://android-review.googlesource.com/#/c/177621/
https://android-review.googlesource.com/#/c/182240/

Change-Id: I8ef055b3a53d3375f07f20ab16289af7f8066aa0
Tracked-On: OAM-33405
Signed-off-by: Nicolae Natea <nicolaex.natea@intel.com>
Reviewed-on: 519448